Change file composition
